requirements.txt
# Django Framework
Django==5.0.7
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (opción para producción)

# Tareas asincrónicas
celery==5.3.6
redis==5.0.1

# Análisis de documentos PDF
pdfminer.six==20231228
PyMuPDF==1.23.14

# Inteligencia Artificial y análisis de texto
spacy==3.7.2
transformers==4.36.2
torch==2.1.2

# APIs y requests
requests==2.31.0

# Validación y formularios
django-crispy-forms==2.1
crispy-tailwind==0.5.0

# Autenticación y seguridad
django-cors-headers==4.3.1
python-decouple==3.8  # Para variables de entorno

# Desarrollo y utilidades
django-extensions==3.2.3
pillow==10.1.0  # Para manejo de imágenes

# Timezone
pytz==2023.3

# Testing
pytest-django==4.8.0
coverage==7.3.4

# Frontend utilities (si decides usar HTMX)
django-htmx==1.17.0

# Calendario y fechas
python-dateutil==2.8.2

# Logging y monitoreo
django-debug-toolbar==4.2.0  # Solo para desarrollo
